requirements.txt
Flask
Flask-SQLAlchemy
spacy
scikit-learn
python-docx
